fix: call cancel when the promise is still pending
After the event sequence is complete, the promise
for controlling the schedule gets resolved and
create a new schedule in the next tick.
Also, the public promise gets resolved too.

The previous step simply cancels all the promises
without looking up its state. This will force the
canceled child event-sequence-listener instance
close the newly created schedule, and create the
rejected promise into the queue of the promise
store. In this case, the next calling of `promise`.`race`
will immediately be rejected because it will get
rejected promise which shouldn't exist in the
current schedule.